+++ zarrs_metadata/src/v3.rs
@@ -10,7 +10,7 @@ pub use group::GroupMetadataV3;
 mod metadata;
 pub use metadata::{
     AdditionalField, AdditionalFields, ConfigurationInvalidError, MetadataConfiguration,
-    MetadataConfigurationSerialize, MetadataV3, UnsupportedAdditionalFieldError,
+    MetadataConfigurationSerialize, MetadataV3,
 };
 
 /// Zarr V3 node metadata ([`ArrayMetadataV3`] or [`GroupMetadataV3`]).